# swagger.yml
basePath: /api/
consumes:
- application/json
- text/vnd.yaml
- application/toml
definitions:
  APIError:
    properties:
      details:
        type: object
        x-go-name: Details
      error:
        $ref: '#/definitions/Error'
    type: object
    x-go-package: github.com/runar-rkmedia/skiver/requestContext
...
  CreatorSource:
    type: string
    x-go-package: github.com/runar-rkmedia/skiver/types
...
  Error:
    properties:
      code:
        $ref: '#/definitions/ErrorCodes'
      error:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/runar-rkmedia/skiver/requestContext
  ErrorCodes:
    type: string
    x-go-package: github.com/runar-rkmedia/skiver/requestContext
...
  LocaleSetting:
    properties:
      auto_translation:
        description: |-
          If set, will allow registered translation-services to translate from other languages to this locale.
          This might help speed up translations for new locales.
          See the Config or Organization-settings for instructions on how to set up translation-services.

          Organization-settings are not yet available.

          TODO: implement organization-settings
        type: boolean
        x-go-name: AutoTranslation
      enabled:
        description: If set, the locale will be visible for editing.
        type: boolean
        x-go-name: Enabled
      publish:
        description: |-
          If set, the associated translations will be published in releases.
          This is useful for when adding new locales, and one don't want to publish it to users until it is complete
        type: boolean
        x-go-name: Publish
    type: object
    x-go-package: github.com/runar-rkmedia/skiver/types
  LocaleSettingInput:
    properties:
      auto_translation:
        description: |-
          If set, will allow registered translation-services to translate from other languages to this locale.
          This might help speed up translations for new locales.
          See the Config or Organization-settings for instructions on how to set up translation-services.

          Organization-settings are not yet available.

          TODO: implement organization-settings
        type: boolean
      enabled:
        description: If set, the locale will be visible for editing.
        type: boolean
      publish:
        description: |-
          If set, the associated translations will be published in releases.
          This is useful for when adding new locales, and one don't want to publish it to users until it is complete
        type: boolean
    type: object
...
  Project:
    properties:
      category_ids:
        items:
          type: string
        type: array
        x-go-name: CategoryIDs
      created_at:
        description: Time of which the entity was created in the database
        format: date-time
        type: string
        x-go-name: CreatedAt
      created_by:
        description: User id refering to the user who created the item
        type: string
        x-go-name: CreatedBy
      deleted:
        description: |-
          If set, the item is considered deleted. The item will normally not get deleted from the database,
          but it may if cleanup is required.
        format: date-time
        type: string
        x-go-name: Deleted
      description:
        type: string
        x-go-name: Description
      id:
        description: Unique identifier of the entity
        type: string
        x-go-name: ID
      included_tags:
        items:
          type: string
        type: array
        x-go-name: IncludedTags
      locales:
        additionalProperties:
          $ref: '#/definitions/LocaleSetting'
        type: object
        x-go-name: LocaleIDs
      short_name:
        type: string
        x-go-name: ShortName
      snapshots:
        additionalProperties:
          $ref: '#/definitions/ProjectSnapshotMeta'
        type: object
        x-go-name: Snapshots
      title:
        type: string
        x-go-name: Title
      updated_at:
        description: Time of which the entity was updated, if any
        format: date-time
        type: string
        x-go-name: UpdatedAt
      updated_by:
        description: User id refering to who created the item
        type: string
        x-go-name: UpdatedBy
    required:
    - created_at
    - id
    type: object
    x-go-package: github.com/runar-rkmedia/skiver/types
...
  ProjectSnapshotMeta:
    properties:
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      created_by:
        type: string
        x-go-name: CreatedBy
      description:
        type: string
        x-go-name: Description
      hash:
        format: uint64
        type: integer
        x-go-name: Hash
      id:
        type: string
        x-go-name: SnapshotID
      uploadMeta:
        items:
          $ref: '#/definitions/UploadMeta'
        type: array
        x-go-name: UploadMeta
    type: object
    x-go-package: github.com/runar-rkmedia/skiver/types
...
  Translation:
    properties:
      aliases:
        items:
          type: string
        type: array
        x-go-name: Aliases
      category:
        type: string
        x-go-name: CategoryID
      created_at:
        description: Time of which the entity was created in the database
        format: date-time
        type: string
        x-go-name: CreatedAt
      created_by:
        description: User id refering to the user who created the item
        type: string
        x-go-name: CreatedBy
      deleted:
        description: |-
          If set, the item is considered deleted. The item will normally not get deleted from the database,
          but it may if cleanup is required.
        format: date-time
        type: string
        x-go-name: Deleted
      description:
        type: string
        x-go-name: Description
      id:
        description: Unique identifier of the entity
        type: string
        x-go-name: ID
      key:
        type: string
        x-go-name: Key
      parent_translation:
        type: string
        x-go-name: ParentTranslationID
      references:
        items:
          type: string
        type: array
        x-go-name: References
      tags:
        items:
          type: string
        type: array
        x-go-name: Tag
      title:
        type: string
        x-go-name: Title
      updated_at:
        description: Time of which the entity was updated, if any
        format: date-time
        type: string
        x-go-name: UpdatedAt
      updated_by:
        description: User id refering to who created the item
        type: string
        x-go-name: UpdatedBy
      value_ids:
        items:
          type: string
        type: array
        x-go-name: ValueIDs
      variables:
        additionalProperties:
          type: object
        type: object
        x-go-name: Variables
    required:
    - created_at
    - id
    type: object
    x-go-package: github.com/runar-rkmedia/skiver/types
  TranslationInput:
    properties:
      category_id:
        maxLength: 100
        minLength: 3
        type: string
      description:
        maxLength: 8000
        minLength: 1
        type: string
      key:
        maxLength: 400
        minLength: 1
        pattern: ^[^\s]*$
        type: string
      title:
        maxLength: 400
        minLength: 1
        type: string
      variables:
        additionalProperties:
          type: object
        description: key/value type. The value can be any type, but the key must a
          string.
        type: object
    required:
    - category_id
    - key
    type: object
  TranslationValue:
    properties:
      context:
        additionalProperties:
          type: string
        type: object
        x-go-name: Context
      created_at:
        description: Time of which the entity was created in the database
        format: date-time
        type: string
        x-go-name: CreatedAt
      created_by:
        description: User id refering to the user who created the item
        type: string
        x-go-name: CreatedBy
      deleted:
        description: |-
          If set, the item is considered deleted. The item will normally not get deleted from the database,
          but it may if cleanup is required.
        format: date-time
        type: string
        x-go-name: Deleted
      id:
        description: Unique identifier of the entity
        type: string
        x-go-name: ID
      locale_id:
        description: locale ID
        type: string
        x-go-name: LocaleID
      source:
        $ref: '#/definitions/CreatorSource'
      translation_id:
        description: Translation ID
        type: string
        x-go-name: TranslationID
      updated_at:
        description: Time of which the entity was updated, if any
        format: date-time
        type: string
        x-go-name: UpdatedAt
      updated_by:
        description: User id refering to who created the item
        type: string
        x-go-name: UpdatedBy
      value:
        description: The pre-interpolated value to use  with translations
        example: The {{productName}} fires up to {{count}} bullets of {{subject}}.
        type: string
        x-go-name: Value
    required:
    - created_at
    - id
    type: object
    x-go-package: github.com/runar-rkmedia/skiver/types
...
  UpdateProjectInput:
    properties:
      description:
        maxLength: 8000
        minLength: 1
        type: string
      id:
        maxLength: 36
        minLength: 3
        type: string
      locales:
        additionalProperties:
          $ref: '#/definitions/LocaleSettingInput'
        type: object
      short_name:
        maxLength: 20
        minLength: 1
        pattern: ^[a-z1-9]*$
        type: string
      title:
        maxLength: 400
        minLength: 1
        type: string
    required:
    - id
    type: object
  UpdateTranslationInput:
    properties:
      description:
        maxLength: 8000
        minLength: 0
        type: string
      id:
        maxLength: 36
        minLength: 3
        type: string
      key:
        maxLength: 400
        minLength: 1
        pattern: ^[^\s]*$
        type: string
      title:
        maxLength: 300
        minLength: 0
        type: string
      variables:
        additionalProperties: true
        type: object
    required:
    - id
    type: object
  UpdateTranslationValueInput:
    properties:
      context_key:
        description: If set, it will add/update the context for that key instead of
          the original value
        maxLength: 100
        minLength: 1
        pattern: ^[^\s]*$
        type: string
      id:
        maxLength: 36
        minLength: 3
        type: string
      value:
        maxLength: 8000
        minLength: 0
        type: string
    required:
    - id
    type: object
  UploadMeta:
    properties:
      id:
        type: string
        x-go-name: ID
      locale:
        type: string
        x-go-name: Locale
      locale_key:
        type: string
        x-go-name: LocaleKey
      parent:
        type: string
        x-go-name: Parent
      provider_id:
        type: string
        x-go-name: ProviderID
      provider_name:
        type: string
        x-go-name: ProviderName
      size:
        format: int64
        type: integer
        x-go-name: Size
      tag:
        type: string
        x-go-name: Tag
      url:
        type: string
        x-go-name: URL
    type: object
    x-go-package: github.com/runar-rkmedia/skiver/types
...
host: localhost:8756
info:
  description: |
    Api fo Skiver-api. <a href="https://insomnia.rest/run/?label=&uri=http://localhost:8756/api/swagger.yml" target="_blank"><img src="https://insomnia.rest/images/run.svg" alt="Run in Insomnia"></a>
    # About formatting
    This api produces output in json, yaml and toml. Every api-route takes parameters for controlling which content-type should be produced.
    In General, you can control this with these parameters:

      - `Accept`-header
      - Appending `.json`/`.yaml`/`.toml` to the path (not all routes supports this.)
      - `out_format`-query-parameter
      - `Content-Type`-header
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  title: Skiver API.
  version: 0.0.1
paths:
...
  /project/:
    get:
      operationId: getProject
      responses:
        "200":
          $ref: '#/responses/ProjectsResponse'
        "500":
          $ref: '#/responses/apiError'
      summary: List projects
      tags:
      - project
    post:
      operationId: createProject
      responses:
        "200":
          $ref: '#/responses/ProjectResponse'
        "500":
          $ref: '#/responses/apiError'
      summary: Create a new project
      tags:
      - project
    put:
      operationId: updateProject
      parameters:
      - in: body
        name: UpdateProject
        required: true
        schema:
          $ref: '#/definitions/UpdateProjectInput'
      responses:
        "200":
          $ref: '#/responses/ProjectResponse'
        "500":
          $ref: '#/responses/apiError'
      summary: Update a project, including its settings
      tags:
      - project
...
  /translation/:
    get:
      operationId: getTranslation
      responses:
        "200":
          $ref: '#/responses/TranslationsResponse'
        "500":
          $ref: '#/responses/apiError'
      summary: List translations
      tags:
      - translation
    post:
      operationId: createTranslation
      parameters:
      - in: body
        name: TranslationInput
        required: true
        schema:
          $ref: '#/definitions/TranslationInput'
      responses:
        "200":
          $ref: '#/responses/TranslationResponse'
        "500":
          $ref: '#/responses/apiError'
      summary: Create a new translation
      tags:
      - translation
    put:
      operationId: updateTranslation
      parameters:
      - in: body
        name: TranslationUpdateInput
        required: true
        schema:
          $ref: '#/definitions/UpdateTranslationInput'
      responses:
        "200":
          $ref: '#/responses/TranslationResponse'
        "500":
          $ref: '#/responses/apiError'
      summary: Update a  translation
      tags:
      - translation
...
  /translationValue/:
    get:
      operationId: getTranslationValue
      responses:
        "200":
          $ref: '#/responses/TranslationValuesResponse'
        "500":
          $ref: '#/responses/apiError'
      summary: List translation-values
      tags:
      - translationValue
    put:
      operationId: updateTranslationValue
      parameters:
      - in: body
        name: UpdateTranslationValueInput
        required: true
        schema:
          $ref: '#/definitions/UpdateTranslationValueInput'
      responses:
        "200":
          $ref: '#/responses/TranslationValueResponse'
        "500":
          $ref: '#/responses/apiError'
      summary: Update a new translation-value for a locale
      tags:
      - translationValue
...
produces:
- application/json
- text/vnd.yaml
- application/toml
responses:
...
  ProjectResponse:
    description: ""
    schema:
      $ref: '#/definitions/Project'
      type: object
  ProjectsResponse:
    description: ""
    schema:
      items:
        $ref: '#/definitions/Project'
      type: array
...
  TranslationResponse:
    description: ""
    schema:
      $ref: '#/definitions/Translation'
      type: object
  TranslationValueResponse:
    description: ""
    schema:
      $ref: '#/definitions/TranslationValue'
      type: object
  TranslationValuesResponse:
    description: ""
    schema:
      items:
        $ref: '#/definitions/TranslationValue'
      type: array
  TranslationsResponse:
    description: ""
    schema:
      items:
        $ref: '#/definitions/Translation'
      type: array
...
  apiError:
    description: ""
    schema:
      $ref: '#/definitions/APIError'
...
schemes:
- http
security:
- cookieAuth: []
swagger: "2.0"
...